requirements.txt
PyQt6==6.6.1
mss==9.0.1
pynput==1.7.6
google-generativeai==0.3.2
python-dotenv==1.0.0
pyinstaller==6.3.0
requests==2.31.0
Pillow==10.1.0
pyperclip==1.8.2
